feat(field-resolver): added field resolver on info-user, product and category while all operations

+++ src/graphql/resolver/product/product.mutation.resolver.ts
@@ -1,13 +1,28 @@
 import productService from '../../../service/product.service';
 const productMutationResolver = {
-  createProduct: async (parent: any, {params}: any, ctx: any) => {
-    return productService.createProduct(params);
+  createProduct: async (
+    parent: any,
+    { params }: any,
+    { isAuthenticate, error, payload }: any
+  ) => {
+    if (!isAuthenticate) throw new Error(error);
+    return productService.createProduct({ ...params, userId: payload.sub });
   },
-  updateProduct: async (parent: any, {params}: any, ctx: any) => {
-    return productService.updateProduct(params);
+  updateProduct: async (
+    parent: any,
+    { params }: any,
+    { isAuthenticate, error, payload }: any
+  ) => {
+    if (!isAuthenticate) throw new Error(error);
+    return productService.updateProduct({...params, userId: payload.sub});
   },
-  deleteProduct: async (parent: any, {params}: any, ctx: any) => {
-    return productService.deleteProduct(params);
+  deleteProduct: async (
+    parent: any,
+    { params }: any,
+    { isAuthenticate, error, payload }: any
+  ) => {
+    if (!isAuthenticate) throw new Error(error);
+    return productService.deleteProduct({...params, userId: payload.sub});
   },
 };
